requirements.txt


fastapi
uvicorn
gunicorn
pandas
scikit-surprise
joblib
numpy<2.0.0
